:recycle: Move Spinner to toolbox
