+++ src/components/toolbox/spinner/spinner.css
@@ -1,4 +1,4 @@
-@import '../app/variables.css';
+@import '../../app/variables.css';
 
 @keyframes rotate {
   0% { transform: rotate(0deg); }
